testConfig: move helpers for parsing yaml to "common.structparse_*"
These helpers are not specific to TestConfig. When we parse YAML (or
structured data that comes from YAML or JSON), then may be useful to
construct similar immutable data classes. Move the code from
"testConfig.py" to "common.py" for reuse.
